#1901 move logic from reviewCR into utils
